Set up Sphinx documentation
I make use of the autodoc and breathe plugins to provide API docs of
the Python and C++ interfaces, respectively. breathe parses output
from Doxygen, which must be run first.

I have provided an outline for general documentation, but have not
written any of it yet, as I do not have the knowledge to do so.